Bulk library lookup.
In Pandora, collecting all the tracks from an album causes the album to
be added to your collection, not the individual tracks. There are often
reasons to know library status of individual tracks, so the client
provides two solutions, depending on the user's needs.

If they need the list of tracks in the collection, there's
library_get_all_track_ids. For each album that appears in the collection,
the client queries the album's track list and adds it to the library. As
a result, it is thorough but slow.

If they have a list of track annotations to check, there's
library_contains_tracks. It produces a dict where the key is the track
ID and the value is a boolean indicating if it's in the library. For each
track, if the ID isn't in the library, it checks if the album its on is.