Filter out playlists where no track has a setVideoId.
There's nothing in the ytmusic library to lookup your own user ID. This
is a problem, since a user may have playlists in their library which
they don't own and thus can't manage, and the only indication is the
author's user ID.

The indicator that a track cannot be managed is lacking a setVideoId
field, as this is the handle for most operations. As such, we filter all
tracks which don't have it, and if the playlist is empty after that (but
didn't start empty) consider it to be unmanageable. Note this is also
useful for the case of individual tracks which may become unavailable
for whatever reason in an otherwise normal user-owned playlist.