Add ytmusicapi "hooks".
There's a bug in ytmusicapi which prevents retrieving the track count.
I've added some code to allow patching it in memory while I wait for my
PR to be merged and released.

It makes sense to take this pseudo-hook structure and use it to extract
and propagate YouTube Music data in other use cases, too.