Use light listing for YTM playlists.
When listing playlists on YouTube Music, no info comes back about their
duration. And the call to retrieve all tracks in order to add up their
lengths if very slow. This can lead to the call taking a while to
fininsh.

So instead, for YTM, we use a light listing, only returning information
contained in that initial playlist enumeration (including "count", which
is hacked in there by a hook). The duration is still computed when you
request the full playlist details.