Handle channel IDs which are actually user IDs.
When links to the artist's YouTube channel are present on its
MusicBrainz page, we use that to grab the channel ID. However, YouTube
is fairly inconsistent with these IDs; sometimes they're channels, and
sometimes they're users. On the YouTube side, this makes no difference.
But on the YTM side, the pages are very different. ytmusicapi expects a
channel ID, and makes a number of assumptions about the response
structure based on that, leading to errors when it's a user ID.

So instead of blindly relying on this ID, we confirm it will work later
by trying to retrieve the artist info with it. If it fails for any
reason, we fall back to search.

This also includes a small tweak to the search to filter out artist
results which have no content for whatever reason.